clean up for mqtt bridge
