+++ mqtt_client.py
@@ -35,8 +35,6 @@ def initialise_clients(cname,mqttclient_log=False,cleansession=True,flags=""):
    client.cname=cname
    client.on_connect= on_connect        #attach function to callback
    client.on_message=on_message        #attach function to callback
-#    if mqttclient_log:
-#       client.on_log=on_log
    return client
 
 def initialise_clients(cname,mqttclient_log=False,cleansession=True,flags="",\
@@ -60,7 +58,7 @@ def on_connect(client, userdata, flags, rc):
    also subscribes to topics
    """
    logging.debug("Connected flags"+str(flags)+"result code "\
-    +str(rc)+"client1_id")
+    +str(rc))
 
    if rc==0:
       client.connected_flag=True #old clients use this
@@ -79,11 +77,14 @@ def on_connect(client, userdata, flags, rc):
      client.bad_connection_flag=True #
      client.bad_count +=1
      client.connected_flag=False #
+
 def on_message(client,userdata, msg):
     topic=msg.topic
+    
     m_decode=str(msg.payload.decode("utf-8","ignore"))
-    message_handler(client,m_decode,topic)
     logging.debug("message received "+ m_decode)
+    
+    message_handler(client,m_decode,topic)
     
 def message_handler(client,msg,topic):
     """ 
@@ -95,8 +96,6 @@ def message_handler(client,msg,topic):
     
     # Generate metadata for the message
     tnow=time.time()
-    #data["time_ms"]=int(tnow*1000)
-    #data["time"]=currentTime()
     data["topic"]=topic
 
     # If msg is a usual json => each of its keys is column
@@ -124,45 +123,9 @@ def message_handler(client,msg,topic):
     save_message(client,data)
     #TODO: enable store changes only
 
-    # if command.options["storechangesonly"]:
-    #     if has_changed(client,topic,msg):
-    #         client.q.put(data) #put messages on queue
-    # else:
-    #     client.q.put(data) #put messages on queue
 def save_message(client,data):
     str_data = json.dumps(data)
     keyname = command.options["message_queue_name"]
     client.redis_conn.lpush(keyname,str_data)
     logging.debug(f"done {str_data} saved to Redis")
     
-def has_changed(client,topic,msg):
-    #print("has changed ",options["testmode"])
-    if topic in client.last_message:
-        if client.last_message[topic]==msg:
-            return False
-    client.last_message[topic]=msg
-    return True
-
-def currentTime():
-    s=time.localtime(tnow)
-
-    year=str(s[0])
-    month=s[1]
-    if month <10:
-        month="0"+str(month)
-    day =s[2]
-    if day<10:
-        day="0"+str(day)
-    hours=s[3]
-    if hours<10:
-        hours="0"+str(hours)
-    m=s[4]
-    if m<10:
-        m="0"+str(m)
-    sec=s[5]
-    if sec<10:
-        sec="0"+str(sec)
-
-    ltime =str(year) + "-" + str(month) + "-" + str(day) + "_" + str(hours)
-    ltime=ltime + ":" + str(m) + ":" + str(sec)
-###
